added L1 loss to GAN model, updated batch generators
